Fix windows bugs, implement showMsg
windows.h and windows.c were beridden with numerous bugs, because I
haven't tested them prior to committing earlier. In this snapshot I've
implemented a showMsg function which uses the windows header to draw a
simple dialog box on the screen. Everything seems to be working
correctly, although much work will still be required to make the entire
system useful and responsive.