Pass array length to snakeCreate, default player
I've mistakenly thought that I could use sizeof(X)/sizeof(X[0]) inside a
function which receives an array, but that is not the case since
functions can only receive pointers, not actual arrays (even the braces
syntax evaluates to a pointer). With that fixed, it's possible to
initialize a default player snake and have it drawn on the screen.