Fix snakeCreate function
The former implementation had a memory leak in the event data was needed
in the already-existing "snakes" pointer instead of appending a new
Snake object. Also, the data would not be copied in that scenario, which
caused some initialization problems (most notably, the SnakeDir variable
would not be initialized which usually resulted in a value of 0, which
corresponds to SNAKE_DIR_UP instead of the dir passed explicitly to the
function).