Add collision death
For now the game just freezes permanently when you hit something. A
proper game over message will be added later.
Also, I've added a new loop in the main function, one that will handle
the initial countdown when starting a new game. For now its only job is
to draw and clean the initial screen.