Add window dialogs for pause and game over
Simplified windows.h a little, by changing it into a simple push & pop
mechanism instead of playing with window pointers. Also, since the
windows can be different sizes, only the last one gets drawn now.
Added window hooks, which are basically pointers to functions that can
be run when a window is popped from the window stack. Using these
functions it's possible to for instance end the game after the player
confirms the "Game Over" dialog which appears after the snake dies.