Add food spawning and eating
Finally, it's possible to play a full game. The current system isn't
perfect, the strategy is randomizing food coordinates until an empty
tile is found and placing the food there. There is no safety checking
for the number of fields available, so if the game stretches out for
long enough for there not to be at least as many empty fields as foods,
it will inevitably hang and crash.

A better way would be to cache and update a list of all empty fields and
pick at random from these. That would both eliminate the instability of
having to reroll dice, and provide an easy and quick way to check if
there are enough empty fields at all. But to make this approach work
nicely I would ideally need to implement an object abstraction that
would keep all coordinate-mutable entities in one place.

Maybe I'll reimplement this the "right way" at some point, but for now
I'm satisfied with this roughly working solution.