Fix food spawn boundaries
The former code had a bit of redundancy in checking available tiles.
Also, it makes more sense to limit the food spawning coords to the size
of the target stage, rather than the first stage in the array (even
though they're the same because the target stage for food is LAYER_WALL,
which is also the first stage).